fix(api): simplify handling of test user
With this patch, we handle the test user by simply setting a cookie as soon as
the user hits the /auth/login endpoint. This is the only bit of special case
code that we need; the cookie is otherwise normal, is validated normally, and
none of the affected routes need any special case code. w00t!